+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changed
+- Convert values to a string representation for easier to read snapshots
+- Advance time by 6 seconds after each run
+
+## Added
+- Deleted, frozen and storage charge events
+
 ## [3.4.1] - 2023-03-24
 
 ## Fixed